Not yet working. Attempting to seperate code into classes
