+++ lib/bank.rb
@@ -1,12 +1,14 @@
+require_relative 'statement'
 # This is the Bank class
-class Bank
-  attr_reader :account, :history, :dates
+class Bank < Statement
+  attr_reader :account, :history, :dates, :statement
 
-  def initialize
+  def initialize(statement = Statement.new)
     @account = 0
     @history = []
     @dates = []
     @transactions = []
+    @statement = statement
   end
 
   def hello
@@ -35,16 +37,7 @@ class Bank
     @transactions
   end
 
-  def statement
-    message_statement = "date || credit || debit || balance\n"
-    @transactions.reverse_each do |tr|
-      # Thought about refactoring this into a ternary operator, but it would have been too long.
-      if tr[:credit].nil?
-        message_statement += "#{tr[:date]} || || #{tr[:debit]}.00 || #{tr[:balance]}.00\n"
-      else
-        message_statement += "#{tr[:date]} || #{tr[:credit]}.00 || || #{tr[:balance]}.00\n"
-      end
-    end
-    message_statement
+  def self.statement
+    statement
   end
 end
